Convert all doubles to decimal

+++ accountant/accountant-app/src/main/kotlin/co/nilin/opex/accountant/app/data/PairFeeResponse.kt
@@ -1,9 +1,11 @@
 package co.nilin.opex.accountant.app.data
 
+import java.math.BigDecimal
+
 data class PairFeeResponse(
     val pair:String,
     val direction: String,
     val userLevel: String,
-    val makerFee: Double,
-    val takerFee: Double
+    val makerFee: BigDecimal,
+    val takerFee: BigDecimal
 )
