+++ wallet/wallet-app/src/main/kotlin/co/nilin/opex/wallet/app/controller/PaymentGatewayController.kt
@@ -31,7 +31,7 @@ class PaymentGatewayController(
         val systemUuid = "1"
         val receiverWalletType = "main"
         val convertedAmount = when (request.currency) {
-            PaymentCurrency.RIALS -> (request.amount / 10).toLong()
+            PaymentCurrency.RIALS -> (request.amount / BigDecimal.valueOf(10)).toLong()
             PaymentCurrency.TOMAN -> request.amount.toLong()
         }
